Apply kubeconfig to Draft debug (also fixes pathing bug)
